Refactoring, review nits   - extract AppEngineRuntime enum as standalone, move to .appengine.ui   - move LibrarySelectorGroup to .libraries.ui (and test) and break cycles   - change description of Java 7 / 8 runtime   - be rid of PageValidator
