+++ plugins/com.google.cloud.tools.eclipse.appengine.libraries.ui.test/META-INF/MANIFEST.MF
@@ -9,6 +9,9 @@ Bundle-Version: 0.1.0.qualifier
 Fragment-Host: com.google.cloud.tools.eclipse.appengine.libraries.ui
 Import-Package: com.google.cloud.tools.eclipse.appengine.libraries.ui,
  com.google.cloud.tools.eclipse.test.util,
+ com.google.cloud.tools.eclipse.test.util.ui,
+ org.eclipse.swtbot.swt.finder.widgets,
+ org.hamcrest;version="1.1.0",
  org.junit;version="4.12.0",
  org.mockito;provider=google;version="1.10.19",
  org.mockito.invocation;provider=google;version="1.10.19",
