+++ plugins/com.google.cloud.tools.eclipse.appengine.libraries.ui/META-INF/MANIFEST.MF
@@ -13,6 +13,7 @@ Import-Package: com.google.cloud.tools.eclipse.appengine.libraries,
  com.google.cloud.tools.eclipse.appengine.libraries.model,
  com.google.cloud.tools.eclipse.appengine.ui,
  com.google.cloud.tools.eclipse.util,
+ com.google.common.annotations;version="15.0.0",
  com.google.common.base,
  org.eclipse.core.resources,
  org.eclipse.core.runtime;version="3.5.0",
@@ -20,7 +21,9 @@ Import-Package: com.google.cloud.tools.eclipse.appengine.libraries,
  org.eclipse.jdt.ui.wizards,
  org.eclipse.jface.dialogs,
  org.eclipse.jface.layout,
+ org.eclipse.jface.viewers,
  org.eclipse.jface.wizard,
  org.eclipse.swt,
+ org.eclipse.swt.events,
  org.eclipse.swt.layout,
  org.eclipse.swt.widgets
