+++ plugins/com.google.cloud.tools.eclipse.appengine.libraries/META-INF/MANIFEST.MF
@@ -16,7 +16,8 @@ Require-Bundle: org.eclipse.equinox.common,
 Export-Package: com.google.cloud.tools.eclipse.appengine.libraries,
  com.google.cloud.tools.eclipse.appengine.libraries.model,
  com.google.cloud.tools.eclipse.appengine.libraries.repository
-Import-Package: com.google.cloud.tools.eclipse.util,
+Import-Package: com.google.cloud.tools.eclipse.appengine.ui,
+ com.google.cloud.tools.eclipse.util,
  com.google.cloud.tools.eclipse.util.io,
  com.google.cloud.tools.eclipse.util.service,
  com.google.cloud.tools.eclipse.util.status,
