+++ plugins/com.google.cloud.tools.eclipse.appengine.libraries/src/com/google/cloud/tools/eclipse/appengine/libraries/ILibraryClasspathContainerResolverService.java
@@ -17,6 +17,7 @@
 package com.google.cloud.tools.eclipse.appengine.libraries;
 
 import com.google.cloud.tools.eclipse.appengine.libraries.model.Library;
+import com.google.cloud.tools.eclipse.appengine.ui.AppEngineRuntime;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IPath;
 import org.eclipse.core.runtime.IProgressMonitor;
@@ -29,29 +30,6 @@ import org.eclipse.jdt.core.IJavaProject;
  * Service interface to resolve {@link LibraryClasspathContainer}s.
  */
 public interface ILibraryClasspathContainerResolverService {
-
-  // extract to a standalone enum?
-  enum AppEngineRuntime {
-    STANDARD_JAVA_7(Messages.getString("appengine.runtimes.java7"), null), // $NON-NLS-1$
-    STANDARD_JAVA_8(Messages.getString("appengine.runtimes.java8"), "java8"); // $NON-NLS-1$
-                                                                              // $NON-NLS-2$
-
-    private String label;
-    private String id;
-
-    AppEngineRuntime(String label, String id) {
-      this.label = label;
-      this.id = id;
-    }
-
-    public String getLabel() {
-      return label;
-    }
-
-    public String getId() {
-      return id;
-    }
-  }
 
   /**
    * Resolves all {@link LibraryClasspathContainer}s found on the classpath of
