+++ plugins/com.google.cloud.tools.eclipse.appengine.libraries/src/com/google/cloud/tools/eclipse/appengine/libraries/messages.properties
@@ -13,6 +13,3 @@ TaskResolveLibraries=Resolving App Engine libraries
 TaskResolveLibrariesError=Could not resolve libraries
 UnexpectedConfigurationElement=Unexpected configuration element with name: {0}. Expected element is {1}.
 adding.app.engine.libraries=Adding App Engine Libraries
-
-appengine.runtimes.java7=Java 7
-appengine.runtimes.java8=Java 8
