+++ plugins/com.google.cloud.tools.eclipse.appengine.newproject/META-INF/MANIFEST.MF
@@ -25,6 +25,7 @@ Import-Package: com.google.cloud.tools.eclipse.appengine.facets,
  com.google.cloud.tools.eclipse.appengine.libraries,
  com.google.cloud.tools.eclipse.appengine.libraries.model,
  com.google.cloud.tools.eclipse.appengine.libraries.repository,
+ com.google.cloud.tools.eclipse.appengine.libraries.ui,
  com.google.cloud.tools.eclipse.appengine.ui,
  com.google.cloud.tools.eclipse.preferences,
  com.google.cloud.tools.eclipse.ui.util,
