+++ plugins/com.google.cloud.tools.eclipse.appengine.ui/META-INF/MANIFEST.MF
@@ -8,9 +8,7 @@ Bundle-Vendor: %providerName
 Bundle-Localization: plugin
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Export-Package: com.google.cloud.tools.eclipse.appengine.ui
-Import-Package: com.google.cloud.tools.eclipse.appengine.libraries,
- com.google.cloud.tools.eclipse.appengine.libraries.model,
- com.google.common.annotations;version="[20.0.0,21.0.0)",
+Import-Package: com.google.common.annotations;version="[20.0.0,21.0.0)",
  com.google.common.base;version="[20.0.0,21.0.0)",
  org.eclipse.core.databinding,
  org.eclipse.core.databinding.conversion,
