+++ plugins/com.google.cloud.tools.eclipse.appengine.ui/src/com/google/cloud/tools/eclipse/appengine/ui/messages.properties
@@ -14,3 +14,6 @@ cloud.sdk.out.of.date=Cloud SDK too old
 appengine.java.component.missing=App Engine Java Component is missing
 appengine.java.component.not.installed=The Cloud SDK App Engine Java component is not installed
 appengine.libraries.group=Libraries to add to build path
+
+appengine.runtimes.java7=App Engine Java 7 runtime
+appengine.runtimes.java8=App Engine Java 8 runtime
